Refactor command package
